# music-library-api/docs/swagger.yaml
basePath: /api
definitions:
  dto.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.PlaylistResponse:
    properties:
      album_cover:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      track_ids:
        items:
          type: string
        type: array
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role_key:
        description: Pass ADMIN_ROLE_KEY to register as admin
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.TrackListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.TrackResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
    type: object
  dto.TrackResponse:
    properties:
      album:
        type: string
      artist:
        type: string
      created_at:
        type: string
      duration:
        type: integer
      file_id:
        type: string
      genre:
        type: string
      id:
        type: string
      release_year:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  dto.UpdateRoleRequest:
    properties:
      role:
        enum:
        - admin
        - artist
        - user
        type: string
    required:
    - role
    type: object
  dto.UpdateTrackRequest:
    properties:
      album:
        type: string
      artist:
        type: string
      genre:
        type: string
      release_year:
        type: integer
      title:
        type: string
    required:
    - artist
    - title
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  description: Clean architecture Music Library API with Golang (Gin) + MongoDB, featuring
    CRUD, MP3 uploads, search, and basic MP3 streaming.
  title: Music Library API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login and receive JWT token
      parameters:
      - description: Login Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register as admin, artist, or user
      parameters:
      - description: Register Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /playlists:
    get:
      consumes:
      - application/json
      description: Retrieve paginated playlists
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get all playlists
      tags:
      - Playlists
    post:
      consumes:
      - multipart/form-data
      description: Create new playlist
      parameters:
      - description: Playlist title
        in: formData
        name: title
        required: true
        type: string
      - description: Album cover image
        in: formData
        name: album_cover
        type: file
      - collectionFormat: csv
        description: Track IDs
        in: formData
        items:
          type: string
        name: track_ids
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PlaylistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create playlist
      tags:
      - Playlists
  /playlists/{id}:
    delete:
      consumes:
      - application/json
      description: Remove playlist by ID
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete playlist
      tags:
      - Playlists
    get:
      consumes:
      - application/json
      description: Get playlist detail
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlaylistResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get playlist by ID
      tags:
      - Playlists
    patch:
      consumes:
      - multipart/form-data
      description: Update playlist by ID (partial update, form-data)
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      - description: Playlist title
        in: formData
        name: title
        type: string
      - description: Album cover image
        in: formData
        name: album_cover
        type: file
      - collectionFormat: csv
        description: Track IDs, comma separated
        in: formData
        items:
          type: string
        name: track_ids
        type: array
      - default: append
        description: Track update mode
        enum:
        - append
        - overwrite
        in: formData
        name: mode
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlaylistResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update playlist
      tags:
      - Playlists
  /playlists/{id}/stream:
    get:
      description: Streams an M3U playlist file containing URLs to the MP3 tracks.
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - audio/x-mpegurl
      responses:
        "200":
          description: M3U playlist content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream playlist as M3U
      tags:
      - Playlists
  /tracks:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of tracks
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrackListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get list of tracks
      tags:
      - tracks
    post:
      consumes:
      - multipart/form-data
      description: Upload a new MP3 track and save metadata
      parameters:
      - description: Artist
        in: formData
        name: artist
        required: true
        type: string
      - description: Album
        in: formData
        name: album
        type: string
      - description: Genre
        in: formData
        name: genre
        type: string
      - description: Release Year
        in: formData
        name: release_year
        type: integer
      - description: MP3 File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TrackResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload a new track
      tags:
      - tracks
  /tracks/{id}:
    delete:
      description: Delete a track by ID
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a track
      tags:
      - tracks
    get:
      description: Retrieve a single track by its ID
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrackResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get track by ID
      tags:
      - tracks
    patch:
      consumes:
      - application/json
      description: Update track metadata by ID
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: string
      - description: Track update info
        in: body
        name: track
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTrackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrackResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a track
      tags:
      - tracks
  /tracks/{id}/stream:
    get:
      description: Stream MP3 file, support Range header
      parameters:
      - description: Track ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - audio/mpeg
      responses:
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream an audio track
      tags:
      - tracks
  /tracks/search:
    get:
      description: Search tracks by query string
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TrackListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search tracks
      tags:
      - tracks
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all users
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - Users
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Update a user's role by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: New role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user role (Admin only)
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your token}" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"